Initial commit
So far, we can have created an AnnData object, with var metadata added.
The goal is to get the average expression value for a list of genes in each cluster. Therefore, we have also created a set of cluster names.

Next steps:
- Filter rows by the value in one column: Cluster_name. That is the first step to getting a list of cellId items that all have the same Cluster_name, meaning that all the cells are part of the same cluster.
- This would allow us to then get a list of genes for each cell and then get the expression value for each gene in our list of interest for each cell.
- Finally, we can get the average expression value for each gene.

Worth considering:
* Maybe it would be smarter to get the average expression value for each gene for a list of cells in the same cluster directly, rather than in two steps.

with '#' will be ignored, and an empty message aborts the commit.