New: - Modified the function getAvgExprForGeneInCluster to return 0 for genes that are not present in the dataset. - Created an instantiation of the solution dataframe in one call, according to --> http://falexwolf.de/blog/171223_AnnData_indexing_views_HDF5-backing/ and --> https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.html (see: Constructing DataFrame from numpy ndarray)
The solution shown for current dataset has many 0.0 columns, but I did verify th
at the number of genes in target list that do not appear in the dataset is the same as the number of columns with only 0.0s.

Todo:
Verify that the data is correct for at least one cluster and gene. Modify code if needed.